Generalize rowchart serializer and query object
