+++ app/models/gobierto_people/interest_group.rb
@@ -7,6 +7,7 @@ module GobiertoPeople
 
     include GobiertoCommon::Sluggable
     include GobiertoCommon::Metadatable
+    include GobiertoCommon::UrlBuildable
 
     belongs_to :site
     has_many :events, class_name: "GobiertoCalendars::Event", dependent: :nullify
@@ -23,10 +24,6 @@ module GobiertoPeople
 
     def attributes_for_slug
       [name]
-    end
-
-    def events_count
-      attributes["events_count"] || events.count
     end
 
   end
