+++ app/queries/gobierto_people/interest_groups_query.rb
@@ -1,24 +1,7 @@
 # frozen_string_literal: true
 
 module GobiertoPeople
-  class InterestGroupsQuery
-
-    attr_reader :limit, :relation
-
-    DEFAULT_LIMIT = 10
-
-    def initialize(params = {})
-      @relation = (params[:relation] || model.all).joins(:events)
-      append_query_conditions(params[:conditions]) if params[:conditions]
-      @limit = params[:limit] || DEFAULT_LIMIT
-    end
-
-    def results
-      relation.select("#{model.table_name}.*, COUNT(*) AS events_count")
-              .group(:id)
-              .order("events_count DESC")
-              .limit(limit)
-    end
+  class InterestGroupsQuery < RowchartItemsQuery
 
     private
 
@@ -43,14 +26,6 @@ module GobiertoPeople
 
     def model
       InterestGroup
-    end
-
-    def events_table
-      ::GobiertoCalendars::Event.table_name
-    end
-
-    def append_condition(attribute_name, attribute_value, operator = "=")
-      @relation = relation.where("#{events_table}.#{attribute_name} #{operator} ?", attribute_value)
     end
 
   end
